+++ src/zc/buildout/patches.py
@@ -15,41 +15,22 @@
 
 def patch_Distribution():
     try:
-        from pkg_resources import _remove_md5_fragment
         from pkg_resources import Distribution
 
-        if hasattr(Distribution, 'location'):
-            return
-
-        # prepare any Distribution built before monkeypatch
-        from pkg_resources import working_set
-        for dist in working_set:
-            dist._location = dist.location
-            dist._location_without_md5 = _remove_md5_fragment(dist.location)
-
         def hashcmp(self):
-            without_md5 = getattr(self, '_location_without_md5', '')
-            return (
-                self.parsed_version,
-                self.precedence,
-                self.key,
-                without_md5,
-                self.py_version or '',
-                self.platform or '',
-            )
+            if hasattr(self, '_hashcmp'):
+                return self._hashcmp
+            else:
+                self._hashcmp = result = (
+                    self.parsed_version,
+                    self.precedence,
+                    self.key,
+                    self.location,
+                    self.py_version or '',
+                    self.platform or '',
+                )
+                return result
 
-        def get_location(self):
-            try:
-                result = self._location
-            except AttributeError:
-                result = ''
-            return result
-
-        def set_location(self, l):
-            self._location = l
-            self._location_without_md5 = _remove_md5_fragment(l)
-
-        setattr(Distribution, 'location', property(get_location, set_location))
         setattr(Distribution, 'hashcmp', property(hashcmp))
     except ImportError:
         return