+++ CHANGES.rst
@@ -4,10 +4,12 @@ Change History
 3.0.0a2 (unreleased)
 ====================
 
-- Warn when a script name passed in 'scripts' argument of easy_install.scripts
+- Ignore `.git` when computing signature of a recipe develop egg
+
+- Warn when the name passed to `zc.recipe.egg:scripts`
   is not defined in egg entry points.
 
-- Show only once pip warning about python version.
+- Show pip warning about Python version only once.
 
 - Better patch for ``pkg_resources.Distribution.hashcmp`` performance.